Refactor landing page components to use centralized imports from the landing library
