+++ apps/demo/tsconfig.app.json
@@ -21,5 +21,5 @@
     "vite.config.ts",
     "vitest.config.ts"
   ],
-  "include": ["src/**/*.js", "src/**/*.jsx", "src/**/*.ts", "src/**/*.tsx", "../../libs/shell/src/lib/nav/NavMain.tsx", "../../libs/shell/src/lib/nav/NavUser.tsx", "../../libs/shell/src/lib/nav/TeamSwitcher.tsx"]
+  "include": ["src/**/*.js", "src/**/*.jsx", "src/**/*.ts", "src/**/*.tsx", "../../libs/shell/src/lib/nav/NavMain.tsx", "../../libs/shell/src/lib/nav/NavUser.tsx", "../../libs/shell/src/lib/nav/TeamSwitcher.tsx", "../../libs/landing/src/lib/components/AboutSection.tsx", "../../libs/landing/src/lib/components/CommunitySection.tsx", "../../libs/landing/src/lib/components/DemoSection.tsx", "../../libs/landing/src/lib/components/FeaturesSection.tsx", "../../libs/landing/src/lib/components/Footer.tsx", "../../libs/landing/src/lib/components/GetStartedSection.tsx", "../../libs/landing/src/lib/components/HeroSection.tsx"]
 }
