+++ libs/landing/src/index.ts
@@ -1,1 +1,1 @@
-export * from './lib/landing';
+export * from './lib/components';
