+++ libs/landing/src/lib/components/AboutSection.tsx
