+++ libs/landing/src/lib/components/CommunitySection.tsx
