+++ libs/landing/src/lib/components/FeaturesSection.tsx
