+++ libs/landing/src/lib/components/Footer.tsx
