+++ libs/landing/src/lib/components/HeroSection.tsx
