+++ libs/landing/src/lib/components/index.ts
@@ -1,0 +1,7 @@
+export * from './AboutSection';
+export * from './CommunitySection';
+export * from './DemoSection';
+export * from './FeaturesSection';
+export * from './HeroSection';
+export * from './Footer';
+export * from './GetStartedSection';
