+++ libs/landing/src/lib/landing.tsx
@@ -1,13 +1,11 @@
-import styled from 'tailwind';
 
-const StyledLanding = styled.div`
-  color: pink;
-`;
+
+
 export function Landing() {
   return (
-    <StyledLanding>
+
       <h1>Welcome to Landing!</h1>
-    </StyledLanding>
+
   );
 }
