+++ libs/shadcnui-blocks/src/lib/shadcnui-blocks.tsx
@@ -1,13 +1,11 @@
-import styled from 'tailwind';
 
-const StyledShadcnuiBlocks = styled.div`
-  color: pink;
-`;
+
+
 export function ShadcnuiBlocks() {
   return (
-    <StyledShadcnuiBlocks>
+
       <h1>Welcome to ShadcnuiBlocks!</h1>
-    </StyledShadcnuiBlocks>
+
   );
 }
